Convert HMFRow
